jdict: infer the stride to avoid some reallocs
If people use jdict with dictionaries they didn't generate they won't
know the stride. We can make a guess then parse the first term bank
to figure out what the stride actually was. Then we can allocate
enough space for all the entries (with some small amount of waste).

This gives an insignificant speed up now but it should be better
if multithreading is added.